Move test of controller into folder controller
